python/odps_data_comparator/requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
pyodps==0.11.6.1
loguru==0.7.2
numpy==1.26.4
openpyxl==3.1.5